Fixed issues with usupported lbl types

+++ TranscodingPredictor/PlaybackModels/PlaybackStatisticsItem.cs
@@ -5,10 +5,10 @@ namespace PlaybackModels
 {
     public class PlaybackStatisticsItem
     {
-        public virtual int Label { get; set; }
+        public virtual float Label { get; set; }
         public virtual long InteractionId { get; set; }
         public virtual int OutputType { get; set; }
-        public virtual long Duration { get; set; }
+        public virtual float Duration { get; set; }
         public virtual int AgentId { get; set; }
     }
 }
